Добавлено тестирование на компиляцию всех классов и всех модулей

+++ backlightinterface.cpp
@@ -5,6 +5,6 @@ BackLightInterface::~BackLightInterface(){}
 bool BackLightInterface::IsOn(){
 	return m_enable;
 }
-uint8_t BackLightInterface::GetPower(){
+unsigned int BackLightInterface::GetPower(){
 	return power;
 }
